requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
gunicorn==21.2.0 